fix: changing rate type and adding columns to make rate average

+++ src/main/java/com/tattoo_marketplace/application/dto/tattoo_artist/TattooArtistExtendedResponse.java
@@ -14,7 +14,7 @@ public record TattooArtistExtendedResponse(
         Integer age,
         LocalDateTime createdAt,
         LocalDateTime updatedAt,
-        Integer rate,
+        Double rate,
         Set<CategoryResponse> categories,
         List<String> images,
         String profilePicture
